Added Mapstruct, refactored class, added create , update, patch, and getByID ticket APIs.

+++ src/main/java/com/caerus/ticketservice/repository/DocumentInfoRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.caerus.ticketservice.domain.DocumentInfo;
 @Repository
-public interface DocumentInfoRepository extends JpaRepository<DocumentInfo, Integer> {}
+public interface DocumentInfoRepository extends JpaRepository<DocumentInfo, Long> {}
