+++ src/main/java/com/caerus/ticketservice/repository/TicketDetailRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.caerus.ticketservice.domain.TicketDetail;
 @Repository
-public interface TicketDetailRepository extends JpaRepository<TicketDetail, Integer> {}
+public interface TicketDetailRepository extends JpaRepository<TicketDetail, Long> {}
