+++ src/main/java/com/caerus/ticketservice/repository/TicketRepository.java
@@ -3,4 +3,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.caerus.ticketservice.domain.Ticket;
 @Repository
-public interface TicketRepository extends JpaRepository<Ticket, Integer> {}
+public interface TicketRepository extends JpaRepository<Ticket, Long> {}
